mved methods to DResolveUtil
